requirements.txt
bert-score
blis==0.2.4
boto3==1.17.90
botocore==1.20.90
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
cymem==2.0.5
Cython==0.29.21
dataclasses==0.8
Flask==1.1.2
gensim==4.0.1
idna==2.10
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==2.6.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
murmurhash==1.0.5
neuralcoref==4.0
nltk==3.6.2
numpy==1.19.5
Pillow==8.2.0
plac==0.9.6
preshed==2.0.1
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytorch-pretrained-bert==0.6.2
pyvi==0.1
regex==2021.4.4
requests==2.25.1
rouge==1.0.0
s3transfer==0.4.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.5.4
sentencepiece==0.1.95
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.1.0
spacy==2.1.3
srsly==1.0.5
tabulate==0.8.9
thinc==7.0.8
threadpoolctl==2.1.0
torch==1.8.1
tqdm==4.32.2
transformers==2.0.0
typing-extensions==3.10.0.0
urllib3==1.26.5
wasabi==0.8.2
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.4.1
flask-cors==3.0.10